refactor GoalsScreen
