+++ src/screens/index.js
@@ -1,7 +1,7 @@
 export { default as TodayScreen } from './TodayScreen'
 export { default as WeekScreen } from './WeekScreen'
 export { default as ActivityDetailScreen} from './ActivityDetailScreen/ActivityDetailScreen'
-export { default as GoalsScreen } from './GoalsScreen/GoalsScreen'
+export { default as GoalsScreen } from './GoalsScreen'
 export { default as GoalScreen} from './GoalScreen/GoalScreen'
 export { default as ActivityFormScreen } from './ActivityFormScreen/ActivityFormScreen'
 export { default as GoalFormScreen } from './GoalFormScreen/GoalFormScreen'
